requirements.txt
matplotlib==3.7.0
matplotlib-inline==0.1.6
numpy==1.24.2
scipy==1.10.1
pandas==1.5.3
fastgb @ git+https://gitlab.in2p3.fr/LISA/fastgb.git@8b243fa56cc91b8911f58c0883db40dff4788db2
lisaconstants==1.3.4
lisaorbits==2.3
plotly==5.18.0
st_pages @ git+https://github.com/blackary/st_pages.git
pillow==10.2.0